v4.6.0 Better custom time interval support. PG14. See CHANGELOG for more details.

+++ sql/procedures/reapply_constraints_proc.sql
@@ -64,7 +64,7 @@ IF v_control_type = 'time' OR (v_control_type = 'id' AND v_epoch <> 'none') THEN
     v_partition_suffix := to_char(v_last_partition_timestamp - (v_partition_interval::interval * (v_optimize_constraint + v_premake + 1) ), v_datetime_string);
 ELSIF v_control_type = 'id' THEN
     SELECT child_start_id INTO v_last_partition_id FROM @extschema@.show_partition_info(v_parent_schema||'.'||v_last_partition, v_partition_interval, p_parent_table);
-    v_partition_suffix := (v_last_partition_id - (v_partition_interval::int * (v_optimize_constraint + v_premake + 1) ))::text; 
+    v_partition_suffix := (v_last_partition_id - (v_partition_interval::bigint * (v_optimize_constraint + v_premake + 1) ))::text; 
 END IF;
 
 v_child_stop := @extschema@.check_name_length(v_parent_tablename, v_partition_suffix, TRUE);
